blog on hyperverse
Hyperverse is a Metaverse for the developers who can use the modular contracts to deploy any web3 dapp. This metaverse is not for the general public. It's introducing APTs as smart contracts as smart modules in Web 3. Smart contract developers are very short supply because it requires very high skills to become a successful Smart-Contact developer. Decentrology team proposed a solution that leveraging the skills of advanced developer to create a hub/marketplace of modules of Smart-Contacts. Each smart contract acts as building block, building blocks and building smart-contracts together.